+++ src/exactly_lib/section_document/element_parsers/token_stream_parser.py
@@ -370,6 +370,15 @@ class TokenParser:
                 return key_handler(key)
         return return_value_if_no_match
 
+    def consume_optional_option(self, option_name: OptionName) -> bool:
+        if self.token_stream.is_null:
+            return False
+        elif matches(option_name, self.token_stream.head.source_string):
+            self.token_stream.consume()
+            return True
+        else:
+            return False
+
     def consume_and_handle_optional_option(self,
                                            return_value_if_no_match: T,
                                            argument_parser: Callable[['TokenParser'], T],