+++ test/exactly_lib_test/test_case_utils/files_matcher/files_condition_containment/symbol_references.py
@@ -4,7 +4,7 @@ from exactly_lib.test_case_utils.files_matcher import parse_files_matcher as sut
 from exactly_lib_test.symbol.test_resources.file_matcher import is_file_matcher_reference_to__ref
 from exactly_lib_test.test_case_utils.file_matcher.test_resources import argument_building as fm_args
 from exactly_lib_test.test_case_utils.files_condition.test_resources import arguments_building as fc_args
-from exactly_lib_test.test_case_utils.files_matcher.test_resources.files_condition import CONTAINS_AND_EQUALS_CASES
+from exactly_lib_test.test_case_utils.files_matcher.test_resources.files_condition import FULL_AND_NON_FULL_CASES
 from exactly_lib_test.test_resources.value_assertions import value_assertion as asrt
 
 
@@ -28,7 +28,7 @@ class Test(unittest.TestCase):
 
         parser = sut.files_matcher_parser()
 
-        for case in CONTAINS_AND_EQUALS_CASES:
+        for case in FULL_AND_NON_FULL_CASES:
             fsm = case.arguments_for_fc(fc_argument)
             with self.subTest(case.name):
                 # ACT #