+++ test/exactly_lib_test/test_case_utils/files_matcher/files_condition_containment/validation_failure.py
@@ -3,7 +3,7 @@ import unittest
 from exactly_lib_test.symbol.test_resources.file_matcher import is_file_matcher_reference_to__ref
 from exactly_lib_test.test_case_utils.file_matcher.test_resources import argument_building as fm_args, validation_cases
 from exactly_lib_test.test_case_utils.files_condition.test_resources import arguments_building as fc_args
-from exactly_lib_test.test_case_utils.files_matcher.test_resources.files_condition import CONTAINS_AND_EQUALS_CASES
+from exactly_lib_test.test_case_utils.files_matcher.test_resources.files_condition import FULL_AND_NON_FULL_CASES
 from exactly_lib_test.test_case_utils.files_matcher.test_resources.integration_check import CHECKER
 from exactly_lib_test.test_resources.value_assertions import value_assertion as asrt
 
@@ -20,7 +20,7 @@ class Test(unittest.TestCase):
             fc_args.FileCondition('file-name',
                                   fm_args.SymbolReferenceWSyntax(fm_symbol_name))
         ])
-        for case in CONTAINS_AND_EQUALS_CASES:
+        for case in FULL_AND_NON_FULL_CASES:
             fsm = case.arguments_for_fc(fc_argument)
             with self.subTest(case.name):
                 # ACT & ASSERT #