+++ test/exactly_lib_test/test_case_utils/files_matcher/files_condition_containment/z_package_suite.py
@@ -2,7 +2,7 @@ import unittest
 
 from exactly_lib_test.test_case_utils.files_matcher.files_condition_containment import invalid_syntax, \
     validation_failure, symbol_references, common_execution_outcome, \
-    specific_for_equals, specific_for_contains
+    specific_for_full, specific_for_non_full
 
 
 def suite() -> unittest.TestSuite:
@@ -11,8 +11,8 @@ def suite() -> unittest.TestSuite:
         symbol_references.suite(),
         validation_failure.suite(),
         common_execution_outcome.suite(),
-        specific_for_equals.suite(),
-        specific_for_contains.suite(),
+        specific_for_full.suite(),
+        specific_for_non_full.suite(),
     ])